refactor: seperated columns and unit test case
